+++ src/processor-queue/__tests__/index.js
@@ -3,7 +3,6 @@
 const path = require("path")
 const { Processor } = require("../index")
 const { Job } = require("../queue")
-const { GooglePubSub } = require("../google-pub-sub")
 
 process.env.TOPIC = "test"
 
@@ -111,45 +110,4 @@ test("failure message should cancel processing", async () => {
   } catch (err) {
     expect(err).toBeDefined()
   }
-})
-
-test("instantiate google pubsub", async () => {
-  const pubSub = await new GooglePubSub({ noSubscription: true })
-  expect(pubSub).toBeInstanceOf(GooglePubSub)
-})
-
-test("size check for google publish", async () => {
-  const pubSub = await new GooglePubSub({ noSubscription: true })
-  const msg = Buffer.from("Hello, World!")
-  pubSub.maxPubSubSize = 10000
-  pubSub.pubSubClient = {
-    topic: () => {
-      return {
-        publish: async msg => {
-          expect(msg).toBe(msg)
-        },
-      }
-    },
-  }
-  pubSub.storageClient = {
-    bucket: () => {
-      return {
-        file: path => {
-          expect(path).toEqual(`event-2345`)
-          return {
-            save: async (data, options) => {
-              expect(Buffer.from(data, "base64").toString()).toEqual(
-                "Hello, World!"
-              )
-              expect(options).toEqual({ resumable: false })
-            },
-          }
-        },
-      }
-    },
-  }
-  pubSub.publish("1234", msg)
-
-  pubSub.maxPubSubSize = 2
-  pubSub.publish("2345", msg)
 })
